backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
axios==0.4.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
django-cors-headers==4.7.0
django-filter==25.1
django-log-viewer==1.1.8
django-simple-log-viewer==1.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-excel==2.5.3
drf-renderer-xlsx==1.0.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
